[SDK-2752] Expand test coverage and finish migration to Pest framework (#552)
* tests: Expand test coverage

* tests: Expand test coverage

* tests: Expand test coverage

* tests: Fix unit tests running in Windows

* tests: Expand test coverage

* fix: handleInvitation() should return a URL

* fix: Improve ConfigurableMixin type validation

* tests: Code style pass

* tests: Address test warning about dynamic class names

* tests: Expand test coverage

* Fix merge conflicts

* tests: Fix test warning; PHP < 8.0 doesn't support `mixed` as a typehint

* change: Rename InMemoryStore to MemoryStore to follow naming convention of other storage mediums.

* tests: Update Psr6StoreTest to PEST format.

* tests: Expand test coverage

* tests: Improvements to toolkit::every() tests

* tests: Fix lack of splat operator support for assoc arrays in PHP < 8.0

* tests: Updates for PHP < 8.0 warnings

* tests: Expand test coverage

* tests: Expand test coverage

* tests: Transition to PEST expectations

* tests: Expand test coverage

* tests: Update test coverage

* Add Pest dependencies

* Adopt expectation API

* Shift cleanup

* tests: Adopt Pest-style exception expectations syntax

* tests: Expand test coverage

* tests: Expand test coverage

* Add `hyperf/event` for PSR-4 unit testing

* tests: Expand test coverage

* tests: Resolve phpstan warnings

+++ src/Auth0.php
@@ -152,6 +152,34 @@ final class Auth0
     }
 
     /**
+     * If invitation parameters are present in the request, handle extraction and return a URL for redirection to Universal Login to accept. Returns null if no invitation parameters were found.
+     *
+     * @param string|null                 $redirectUrl Optional. URI to return to after logging out. Defaults to the SDK's configured redirectUri.
+     * @param array<int|string|null>|null $params Additional parameters to include with the request.
+     *
+     * @throws \Auth0\SDK\Exception\ConfigurationException When a Client ID is not configured.
+     * @throws \Auth0\SDK\Exception\ConfigurationException When `redirectUri` is not specified, and supplied SdkConfiguration does not have a default redirectUri configured.
+     *
+     * @link https://auth0.com/docs/universal-login/new-experience
+     * @link https://auth0.com/docs/api/authentication#login
+     */
+    public function handleInvitation(
+        ?string $redirectUrl = null,
+        ?array $params = null
+    ): ?string {
+        $invite = $this->getInvitationParameters();
+
+        if ($invite !== null) {
+            return $this->login($redirectUrl, Toolkit::merge([
+                'invitation' => (string) $invite->invitation,
+                'organization' => (string) $invite->organization,
+            ], $params));
+        }
+
+        return null;
+    }
+
+    /**
      * Delete any persistent data and clear out all stored properties, and return the URI to Auth0 /logout endpoint for redirection.
      *
      * @param string|null                 $returnUri Optional. URI to return to after logging out. Defaults to the SDK's configured redirectUri.
@@ -173,16 +201,19 @@ final class Auth0
 
     /**
      * Delete any persistent data and clear out all stored properties.
+     *
+     * @oaram bool $transient When true, data in transient storage is also cleared.
      */
-    public function clear(): self
-    {
+    public function clear(
+        bool $transient = true
+    ): self {
         // Delete all data in the session storage medium.
         if ($this->configuration()->hasSessionStorage()) {
             $this->configuration->getSessionStorage()->purge();
         }
 
         // Delete all data in the transient storage medium.
-        if ($this->configuration()->hasTransientStorage()) {
+        if ($this->configuration()->hasTransientStorage() && $transient === true) {
             $this->configuration->getTransientStorage()->purge();
         }
 
@@ -273,6 +304,8 @@ final class Auth0
             return false;
         }
 
+        $this->clear(false);
+
         if ($state === null || ! $this->getTransientStore()->verify('state', $state)) {
             $this->clear();
             throw \Auth0\SDK\Exception\StateException::invalidState();
@@ -285,10 +318,6 @@ final class Auth0
                 $this->clear();
                 throw \Auth0\SDK\Exception\StateException::missingCodeVerifier();
             }
-        }
-
-        if ($this->getState()->hasUser()) {
-            $this->clear();
         }
 
         $response = $this->authentication()->codeExchange($code, $redirectUri, $codeVerifier);
@@ -622,18 +651,20 @@ final class Auth0
      * Get the specified parameter from POST or GET, depending on configured response mode.
      *
      * @param string $parameterName Name of the parameter to pull from the request.
+     * @param int $filter Defaults to FILTER_SANITIZE_STRING. The type of PHP filter_var() filter to apply.
      */
     public function getRequestParameter(
-        string $parameterName
+        string $parameterName,
+        int $filter = FILTER_SANITIZE_STRING
     ): ?string {
         $responseMode = $this->configuration()->getResponseMode();
 
         if ($responseMode === 'query' && isset($_GET[$parameterName])) {
-            return filter_var($_GET[$parameterName], FILTER_SANITIZE_STRING, FILTER_NULL_ON_FAILURE);
+            return filter_var($_GET[$parameterName], $filter, FILTER_NULL_ON_FAILURE);
         }
 
         if ($responseMode === 'form_post' && isset($_POST[$parameterName])) {
-            return filter_var($_POST[$parameterName], FILTER_SANITIZE_STRING, FILTER_NULL_ON_FAILURE);
+            return filter_var($_POST[$parameterName], $filter, FILTER_NULL_ON_FAILURE);
         }
 
         return null;
@@ -657,23 +688,6 @@ final class Auth0
         }
 
         return null;
-    }
-
-    /**
-     * If invitation parameters are present in the request, handle extraction and automatically redirect to Universal Login.
-     */
-    public function handleInvitation(): self
-    {
-        $invite = $this->getInvitationParameters();
-
-        if ($invite !== null) {
-            $this->login(null, [
-                'invitation' => (string) $invite->invitation,
-                'organization' => (string) $invite->organization,
-            ]);
-        }
-
-        return $this;
     }
 
     /**
